IOS-4246: Make horizontal scroll animations using a dedicated factory
